improved handling of message merge: no long time-stamp-changed-only commits for otherwise unchanged .po files
git-svn-id: https://kradio.svn.sourceforge.net/svnroot/kradio/trunk@858 1c7eebff-f921-0410-8e14-d196b13b6ee1

+++ merge-messages.sh
@@ -13,8 +13,12 @@ for cat in $catalogs; do
     for pot in $pots ; do
         echo "merging $pot into $cat"
         test -e $cat || cat $pot | sed 's/charset=CHARSET/charset=UTF-8/; s/^#, fuzzy/#/' > $cat
-        msgmerge -q -o $cat.new $cat $pot
-        mv $cat.new $cat
+        if test -e $cat ; then 
+            msgmerge --update -q $cat $pot
+        else
+            msgmerge --update -q -o $cat.new $cat $pot
+            mv $cat.new $cat
+        fi
     done
 done
 echo "Done merging translations"
